update from master

+++ src/shared/engine/AccusationCommand.cpp
@@ -16,7 +16,7 @@ namespace engine{
 
     void AccusationCommand::execute () {
 
-        for (int i=0; i<envelope.size();i++) {
+        for (int i = 0; i < (int)envelope.size();i++) {
             if (accusation.suspect == static_cast<state::SuspectCard&>(envelope.at(0)).getSuspectName() and
                 accusation.weapon == static_cast<state::WeaponCard&>(envelope.at(1)).getWeaponName() and
                 accusation.room == static_cast<state::RoomCard&>(envelope.at(2)).getRoomName()){
